Be able to use different versions for sponge and forge

+++ launch.sh
@@ -2,12 +2,12 @@
 
 set -e
 
-SPONGE_VERSION="${MC_VERSION}-${FORGE_BUILD}-${SPONGE_API_VERSION}-BETA-${SPONGE_BUILD}"
 SPONGE_DOWNLOAD_URL="https://repo.spongepowered.org/maven/org/spongepowered/spongeforge/${SPONGE_VERSION}/spongeforge-${SPONGE_VERSION}.jar"
 
-FORGE_DOWNLOAD_URL="http://files.minecraftforge.net/maven/net/minecraftforge/forge/${MC_VERSION}-${FORGE_VERSION}.${FORGE_BUILD}/forge-${MC_VERSION}-${FORGE_VERSION}.${FORGE_BUILD}-installer.jar"
+FORGE_DOWNLOAD_URL="http://files.minecraftforge.net/maven/net/minecraftforge/forge/${MC_VERSION}-${FORGE_VERSION}/forge-${MC_VERSION}-${FORGE_VERSION}-installer.jar"
 
-SERVER="forge-${MC_VERSION}-${FORGE_VERSION}.${FORGE_BUILD}-universal.jar"
+SERVER="forge-${MC_VERSION}-${FORGE_VERSION}-universal.jar"
+SPONGE="spongeforge-${SPONGE_VERSION}.jar"
 
 if [ ! -f "/minecraft/${SERVER}" ]; then
     echo "Installing server, take a coffee"
@@ -16,9 +16,6 @@ if [ ! -f "/minecraft/${SERVER}" ]; then
 
     cd /tmp/mc
 
-    echo "Downloading sponge ${SPONGE_DOWNLOAD_URL}"
-    curl -s "${SPONGE_DOWNLOAD_URL}" -O
-    
     echo "Downloading forge ${FORGE_DOWNLOAD_URL}"
     curl -s "${FORGE_DOWNLOAD_URL}" -O
 
@@ -29,11 +26,28 @@ if [ ! -f "/minecraft/${SERVER}" ]; then
         echo "eula=$EULA" >> eula.txt
     fi
 
-    java -jar "/tmp/mc/forge-${MC_VERSION}-${FORGE_VERSION}.${FORGE_BUILD}-installer.jar" --installServer
+    java -jar "/tmp/mc/forge-${MC_VERSION}-${FORGE_VERSION}-installer.jar" --installServer
 
-    mkdir -p mods
+    mkdir -p /minecraft/mods
 
-    mv "/tmp/mc/spongeforge-${SPONGE_VERSION}.jar" mods/
+    cd /minecraft
+
+    rm -r /tmp/mc
+fi
+
+if [ ! -f "/minecraft/mods/${SPONGE}" ]; then
+    mkdir -p /tmp/mc
+
+    cd /tmp/mc
+    
+    echo "Downloading sponge ${SPONGE_DOWNLOAD_URL}"
+    curl -s "${SPONGE_DOWNLOAD_URL}" -O
+
+    rm -f "/minecraft/mods/spongeforge-*.jar"
+
+    mv "/tmp/mc/${SPONGE}" /minecraft/mods/
+
+    cd /minecraft
 
     rm -r /tmp/mc
 fi
